backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-Session==0.8.0
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.4
WTForms==3.1.2
